+++ CHANGELOG.md
@@ -3,6 +3,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog].
 
+## [v4.1.6-1.19.1] - 2022-08-06
+### Changed
+- Made network manager check messages are being sent to the correct side on Fabric
+
+## [v4.1.5-1.19.1] - 2022-08-06
+### Fixed
+- Fix type constraints when registering network messages
+
 ## [v4.1.4-1.19.1] - 2022-08-01
 ### Fixed
 - Fixed client configs causing dedicated servers to fail on start-up